Add files via upload
This was the first assignment in the class. 
This was a program written in C to showcase the use of a user input menu. This was a good review of the lower level course in C to make sure everyone was up to speed on the basics like datatypes, structure and standard libraries. 
In this assignment I made a bank menu for a very basic online bank. The numbers were only stored for the session and it was limited to 100 elements on stack array. I try to write self explanatory code using common variable names and reserve comments for logic explanation. The functions manipulate the data and Looking back, there are many ways to improve this code, but It shows how much we learned.