Add files via upload
This was our second assignment.
For this assignment we introduced the concept of a 2D array. This was my first opportunity to implement a 2d array and it took some adjusting to learn to manipulate these arrays. I really enjoyed figuring out how to sort the array, I only used a simple bubble sort, but was happy with the final code. Again, Looking back, I see thing that could be improved and feel like Ive come a long way.