Add files via upload
This was the 3rd assignment we did. This program takes a string in C including spaces until the return char is entered. it then manipulates the string. This really highlighted how to display to the console and how to manipulate and reorder information. Strings are just arrays of ints with an ASCII mask, so I feel like the reason for this assignment was about manipulation of array data rather than focusing on the string aspect.