Add files via upload
This Introduced the use of Structs in C. up until this point, I didn't know what Structs were or how they were used, and once the idea was introduced, I couldn't believe I was missing such a powerful tool. I learned to used and manipulate Structs and became much better at using pointers. This was a very powerful lesson when it comes to expanded utility.