Add files via upload
This assignment was a lot of fun. We built a "gambling" game. its like going to the racetrack where no pups are harmed in the making. This was the first time I implemented Random(). I learned that random isn't always random and how to pair it with time.h to make it more random. This project is where I really started using pointers and building programs that were a little more fluid. The pointers and referencing parts of a struct was also new. Pointers were another one of those powerful tools that I don't really understand how I got around using them.