Add files via upload
For this project, we were given free reign to make whatever we wanted within the program specifications. This was much harder than I thought it would be. It was like when you sit down in a group of people and someone says "Allen's really funny! Allen, you should tell a joke to everyone right now!" then you sit there awkwardly for 15 seconds as every joke you've ever known in your life seems like it never existed. Finally I decided to do a random restaurant selector that I have actually used with my wife when deciding where to go eat. It selects either a random type of food or picks randomly from eatery that you enter as options. I added the $ counter because I was originally going to scale based on price, but that extra feature didn't fit into the sprint.