Add files via upload
This week we learned about linked lists and how to add and remove items. Linked lists seem like a great tool for databases where you'll have lots of dynamic data. We had a study group on Discord and I'm glad we did, because we were not provided reading material for this one and learning Nodes and linked lists from google was much easier working with a team of 3-4 students. I enjoyed working along side the people in our Discord group and learning with them. They were all fantastic and everyone helped each other. I have been a solo learner for a year or two and I noticed the power of having a fresh set of eyes look over something that has you stuck. We also helped each other ask better questions which is probably the most powerful tool I've learned this year. This was the last task before the final and I plan to move on to JAVA next semester to see what OOP is all about.